Added diamond goggle effect
+Diamond goggle lens now opens a detailed crafter window when pressing the stability control key with an empty hand
*Fixed some blocks not ackowledging the copshowium axle as an axle
*Fixed rift beams not changing skulls to wither skulls on the client
*Some staff and goggle beams are a little bit more nicely aligned now
*The less than axis now acts as a Math.min, and greater than acts as Math.max
*All the math master axes have been merged code-wise into AbstractMathAxisTE. This is purely an internal change
*Fixed a bug with some math axes not liking being rotated via wrench

+++ src/main/java/com/Da_Technomancer/crossroads/API/technomancy/LooseBeamRenderable.java
@@ -7,7 +7,7 @@ public class LooseBeamRenderable{
 	public final double x;
 	public final double y;
 	public final double z;
-	public final int length;
+	public final double length;
 	public final float angleX;
 	public final float angleY;
 	public final byte width;
@@ -15,7 +15,7 @@ public class LooseBeamRenderable{
 	public byte lifeTime = 6;
 	public long lastTick = 0;
 	
-	public LooseBeamRenderable(double x, double y, double z, int length, float angleX, float angleY, byte width, int color){
+	public LooseBeamRenderable(double x, double y, double z, double length, float angleX, float angleY, byte width, int color){
 		this.x = x;
 		this.y = y;
 		this.z = z;
@@ -30,7 +30,7 @@ public class LooseBeamRenderable{
 		nbt.setDouble("x", x);
 		nbt.setDouble("y", y);
 		nbt.setDouble("z", z);
-		nbt.setInteger("length", length);
+		nbt.setDouble("length", length);
 		nbt.setFloat("angleX", angleX);
 		nbt.setFloat("angleY", angleY);
 		nbt.setByte("width", width);
@@ -38,6 +38,6 @@ public class LooseBeamRenderable{
 	}
 	
 	public static LooseBeamRenderable readFromNBT(NBTTagCompound nbt){
-		return new LooseBeamRenderable(nbt.getFloat("x"), nbt.getFloat("y"), nbt.getFloat("z"), nbt.getInteger("length"), nbt.getFloat("angleX"), nbt.getFloat("angleY"), nbt.getByte("width"), nbt.getInteger("color"));
+		return new LooseBeamRenderable(nbt.getFloat("x"), nbt.getFloat("y"), nbt.getFloat("z"), nbt.getDouble("length"), nbt.getFloat("angleX"), nbt.getFloat("angleY"), nbt.getByte("width"), nbt.getInteger("color"));
 	}
 }
